# quickcart-backend-enterprise/openapi.yaml
openapi: 3.0.3
info:
  title: QuickCart Backend — Enterprise
  version: "1.0"
  description: >
    Multi-tenant API for QuickCart. Public endpoints (products, orders, payments, devices)
    require an X-Organization-Slug header. Admin endpoints require a Bearer token from
    /auth/login and are further restricted by role (owner / manager / support).
    This spec covers the shape of every endpoint; it is not exhaustively validated against
    the live server — treat it as documentation, not a contract test.
servers:
  - url: http://localhost:4000/api/v1
security:
  - bearerAuth: []
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
...
paths:
  /auth/login:
    post:
      summary: Log in and receive an access + refresh token
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [organizationSlug, email, password]
              properties:
                organizationSlug: { type: string }
                email: { type: string, format: email }
                password: { type: string }
      responses:
        "200": { description: OK }
        "401": { description: Incorrect email or password }
  /auth/refresh:
    post:
      summary: Exchange a refresh token for a new access token
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [refreshToken]
              properties: { refreshToken: { type: string } }
      responses:
        "200": { description: OK }
        "401": { description: Invalid or expired refresh token }
...